Separate slot file
